Blog_Generating_gemini/requirements.txt
streamlit
langchain
google-generativeai
textwrap
